Learn Java in an Hour Workshop - 10/25/16
As Technology chair for my engineering sorority that barely has any CS majors, my responsibilities included maintaining and updating our website. I noticed that most sisters had a bad perception of programming from the minimal exposure they had had. To change this perspective, I designed and implemented an interactive workshop on Beginner Level Java. The object oriented programming aspect of Java is invaluable, and I wrote and taught code on what the sisters could relate to– ­the sorority we are in! I chose to build a class AlphaOmegaEpsilon in order to make coding a fun experience for them. Through a presentation and codepad.io, an online website to code, I had fun teaching them about classes, objects, fields, and methods by relating it to the sorority. Not only did I build a class that illustrated our sorority in code, but also a foundation, an introduction, and a path to further interest in code for them, which brought great contentment to me as they requested